+++ src/pacx/clients/__init__.py
@@ -1,6 +1,7 @@
 from .authorization import AuthorizationRbacClient as AuthorizationRbacClient
 from .connectors import ConnectorsClient as ConnectorsClient
 from .dataverse import DataverseClient as DataverseClient
+from .policy import DataLossPreventionClient as DataLossPreventionClient
 from .governance import GovernanceClient as GovernanceClient
 from .licensing import LicensingClient as LicensingClient
 from .power_platform import PowerPlatformClient as PowerPlatformClient
@@ -15,6 +16,7 @@ __all__ = [
     "AuthorizationRbacClient",
     "ConnectorsClient",
     "DataverseClient",
+    "DataLossPreventionClient",
     "GovernanceClient",
     "LicensingClient",
     "PowerPlatformClient",